Moving tests to new format
