--- elide-async/src/test/resources/META-INF/persistence.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2020, Yahoo Inc.
-  ~ Licensed under the Apache License, Version 2.0
-  ~ See LICENSE file in project root for terms.
-  -->
-
-<persistence xmlns="http://java.sun.com/xml/ns/persistence"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd"
-             version="2.0">
-    <persistence-unit name="asyncStore">
-        <class>com.yahoo.elide.async.models.AsyncQuery</class>
-        <class>com.yahoo.elide.async.models.AsyncQueryResult</class>
-        <class>com.yahoo.elide.async.integration.models.Player</class>
-
-        <properties>
-            <property name="hibernate.dialect" value="org.hibernate.dialect.H2Dialect"/>
-            <property name="hibernate.archive.autodetection" value="" />
-            <property name="javax.persistence.jdbc.driver" value="org.h2.Driver"/>
-            <property name="javax.persistence.jdbc.url" value="jdbc:h2:mem:db1;DB_CLOSE_DELAY=-1;MVCC=TRUE;INIT=RUNSCRIPT FROM 'classpath:create_tables.sql'"/>
-            <property name="javax.persistence.jdbc.user" value=""/>
-            <property name="javax.persistence.jdbc.password" value=""/>
-        </properties>
-    </persistence-unit>
-</persistence>
