--- elide-async/src/test/resources/player.csv
@@ -1,4 +1,0 @@
-id,name
-1,Jon Doe
-2,Jane Doe
-3,Han
