+++ elide-datastore/elide-datastore-hibernate5/src/test/java/com/yahoo/elide/datastores/hibernate5/HibernateDataStoreHarness.java
@@ -5,6 +5,7 @@
  */
 package com.yahoo.elide.datastores.hibernate5;
 
+import com.yahoo.elide.async.models.AsyncQuery;
 import com.yahoo.elide.core.DataStore;
 import com.yahoo.elide.core.datastore.test.DataStoreTestHarness;
 import com.yahoo.elide.utils.ClassScanner;
@@ -53,6 +54,8 @@ public class HibernateDataStoreHarness implements DataStoreTestHarness {
                     .forEach(metadataSources::addAnnotatedClass);
             ClassScanner.getAnnotatedClasses(BookV2.class.getPackage(), Entity.class)
                     .forEach(metadataSources::addAnnotatedClass);
+            ClassScanner.getAnnotatedClasses(AsyncQuery.class.getPackage(), Entity.class)
+            .forEach(metadataSources::addAnnotatedClass);
         } catch (MappingException e) {
             throw new IllegalStateException(e);
         }
