+++ elide-integration-tests/src/test/java/com/yahoo/elide/async/integration/tests/framework/AsyncTestSecurityContext.java
@@ -3,7 +3,8 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.async.integration.framework;
+package com.yahoo.elide.async.integration.tests.framework;
+
 
 import java.security.Principal;
